+++ tests/pna/pixeldataset/snapshots/test_pixel_data_viewer/test_read_layouts_add_marker_counts/read_view_layouts.csv
@@ -1,0 +1,35 @@
+MarkerA,MarkerB,MarkerC,component,index,layout,pixel_type,projection,sample,x,x_norm,y,y_norm,z,z_norm
+0,0,1,3770519d30f36d18,17533628619506136151,wpmds_3d,A,full,test_sample,11847.637205736812,1.0,1.080301475758514e-12,9.118286262474471e-17,-1.6055968053242144e-12,-1.3552042297064593e-16
+1,0,0,3770519d30f36d18,2680297880380432217,wpmds_3d,A,full,test_sample,-8701.345998372599,-0.9514937069516293,-2547.699687924789,-0.2785914065153192,-1193.9848473143593,-0.13056245190430218
+0,1,0,3770519d30f36d18,9914041027802204156,wpmds_3d,A,full,test_sample,191.14716126502702,0.06993630467107738,2419.2095351654652,0.8851325543878907,-1257.400311412728,-0.4600535560689885
+0,0,1,3770519d30f36d18,11721705052923611698,wpmds_3d,B,full,test_sample,191.1471612650279,0.06993630467107764,-2419.2095351654652,-0.88513255438789,1257.4003114127327,0.4600535560689899
+1,0,0,3770519d30f36d18,4489093624312168814,wpmds_3d,B,full,test_sample,5172.760468478328,1.0,5.600614486509197e-13,1.0827128997443661e-16,-2.15801871306354e-13,-4.1718898955674334e-17
+1,0,0,3770519d30f36d18,3527325133011089308,wpmds_3d,B,full,test_sample,-8701.345998372599,-0.9514937069516293,2547.699687924787,0.278591406515319,1193.9848473143566,0.1305624519043019
+0,1,0,4920229146151c29,4419594517623728635,wpmds_3d,A,full,test_sample,5537.222746495467,0.21323318485925494,-25370.69462190507,-0.9770013351448291,-1.3806556648044792e-11,-5.31677373438804e-16
+0,0,1,4920229146151c29,1645819806523959612,wpmds_3d,A,full,test_sample,4469.338884876907,0.589867399062981,5975.3187068685465,0.7886279816729537,1315.0230796268293,0.1735579385828186
+1,0,0,4920229146151c29,1657631243467327470,wpmds_3d,B,full,test_sample,3401.3177291693723,0.5620601598177587,5005.189283082019,0.8270963527580301,5.868470930239122e-12,9.697517172973942e-16
+0,0,1,4920229146151c29,10605176534183593997,wpmds_3d,B,full,test_sample,3473.40523985133,0.22562893483917904,-14997.356773442287,-0.9742133153284949,1.118247262821507e-12,7.264022518951562e-17
+0,0,1,4920229146151c29,6006701602935914176,wpmds_3d,B,full,test_sample,2007.5463585227521,0.5240940784852441,3262.291557420829,0.8516603764979926,-2.1469314125227105e-11,-5.604822202188569e-15
+0,1,0,4920229146151c29,17093482914600361331,wpmds_3d,A,full,test_sample,4469.3388848769055,0.2720423414565224,5975.318706868601,0.36370920483696206,-14636.502763371134,-0.890903237603118
+0,0,1,4920229146151c29,10162086509614058846,wpmds_3d,A,full,test_sample,4469.338884876912,0.2830161492150704,5975.31870686859,0.3783807257205328,-13917.687351378181,-0.8813228044744309
+0,0,1,4920229146151c29,14716292412758382347,wpmds_3d,A,full,test_sample,-36511.25144083845,-0.9997742567485008,-775.9301791678419,-0.021247012566065183,5.131846583986852e-12,1.4052347979817603e-16
+1,0,0,4920229146151c29,12237051952843152705,wpmds_3d,A,full,test_sample,2007.5463585227521,0.5240940784852441,3262.291557420829,0.8516603764979926,-1.965032472168125e-11,-5.129953180520104e-15
+0,0,1,4920229146151c29,17336050916343892849,wpmds_3d,A,full,test_sample,4469.338884876903,0.2874385989934555,5975.31870686857,0.3842933556579014,13641.379283289765,0.8773241525262307
+1,0,0,4920229146151c29,17581986263961994225,wpmds_3d,A,full,test_sample,4469.338884876903,0.2881470633099378,5975.318706868572,0.3852405427458509,13597.787751832804,0.8766761056004326
+0,0,1,4920229146151c29,8754864093431251381,wpmds_3d,B,full,test_sample,-2262.4814161077406,-0.9933423365034062,-262.3843577513568,-0.11519983728267083,6.9162132160265245e-12,3.0365630174247427e-15
+1,0,0,e7d82bca9694eea7,17578746049728489641,wpmds_3d,B,full,test_sample,-1307.1533288881922,-0.6373520311371395,-1426.8850707674014,-0.6957317691463053,-679.4054574869189,-0.3312698202461478
+1,0,0,e7d82bca9694eea7,4633887202036215820,wpmds_3d,A,full,test_sample,-1307.153328888195,-0.6373520311371405,1426.8850707674003,0.6957317691463044,679.405457486919,0.3312698202461477
+1,0,0,e7d82bca9694eea7,3093013882117452616,wpmds_3d,A,full,test_sample,3080.694484075075,0.8645940110835535,-1696.7818282472886,-0.4762002251120523,571.3365667567102,0.16034507040031842
+0,1,0,e7d82bca9694eea7,5828273475010330184,wpmds_3d,B,full,test_sample,3080.6944840750725,0.864594011083553,1696.781828247291,0.4762002251120531,-571.3365667567098,-0.16034507040031834
+0,1,0,e7d82bca9694eea7,6367193580528650492,wpmds_3d,B,full,test_sample,-3547.082310373761,-1.0,-1.1996892813322442e-12,-3.3821861923633546e-16,-4.27687781173581e-13,-1.2057452964166341e-16
+0,0,1,fc07dea9b679aca7,16718381540940362211,wpmds_3d,A,full,test_sample,-2861.85451857026,-0.9916930549679929,-371.19503092741616,-0.12862692069799933,-2.5047098349941335e-12,-8.679348764781484e-16
+1,0,0,fc07dea9b679aca7,3120322361086630706,wpmds_3d,B,full,test_sample,3187.5743884874346,0.3212923200903581,9395.08732530884,0.9469800658149858,-1.239236243601857e-13,-1.249091124853196e-17
+0,1,0,fc07dea9b679aca7,4765690112321800547,wpmds_3d,B,full,test_sample,4935.628902197851,0.582239263222056,-6891.930653090895,-0.8130174908097841,-7.679290072044929e-13,-9.058995889307757e-17
+0,1,0,fc07dea9b679aca7,13496407243000087834,wpmds_3d,B,full,test_sample,2133.5874358176675,0.7475060532761306,1895.9660706481804,0.6642549964550832,-2.9046127500186144e-13,-1.0176361074369819e-16
+0,1,0,fc07dea9b679aca7,10575684261142813337,wpmds_3d,B,full,test_sample,7426.807763912587,0.27403716588160526,-12557.142479547003,-0.4633381994061097,-22839.65689592191,-0.8427463109896606
+1,0,0,fc07dea9b679aca7,3381206478569230700,wpmds_3d,B,full,test_sample,2384.755787235289,0.809671515038575,1728.5694023968067,0.5868833254873906,1.3439328321995829e-12,4.562916413418725e-16
+0,1,0,fc07dea9b679aca7,11449696640771709155,wpmds_3d,B,full,test_sample,2384.7557872352872,0.8096715150385743,1728.5694023968094,0.5868833254873915,-2.0309220459809483e-14,-6.8953799743201794e-18
+0,0,1,fc07dea9b679aca7,1906719288004785482,wpmds_3d,B,full,test_sample,-42106.18027696722,-0.9998248284767882,-788.2240094915599,-0.01871663328060848,-2.308470835836884e-12,-5.4815384399682325e-17
+0,1,0,fc07dea9b679aca7,2986818696398050493,wpmds_3d,B,full,test_sample,7661.309202826124,0.24010620542174582,30974.584931400263,0.9707466250870874,2.720178467108614e-12,8.525066832787033e-17
+0,0,1,fc07dea9b679aca7,5412574254597463076,wpmds_3d,B,full,test_sample,7426.807763912605,0.49003392902216825,-12557.142479547034,-0.8285424993014557,-4105.602932460988,-0.2708949524417054
+0,0,1,fc07dea9b679aca7,11601060255102815255,wpmds_3d,B,full,test_sample,7426.807763912588,0.24237938281590468,-12557.14247954701,-0.409811663486569,26945.259828382903,0.8793785506004523
